ScaleCritiqueClient improvements (#1633)
